refactor: move more stuff into manager.tl
